Percentage of pronouns
This function will eventually identify the pronouns in each text and then return their percentage (how much of the text is made up of these particular parts of speech). As of right now, the code included is for definite and indefinite articles. This is what I'll be modifying to focus on pronouns instead.